fix: group mode split logic (1/N) + compact UI layout
Backend:
- Enable float quantities in models for fractional splits (0.33, 0.5, etc.)
- Update assignment handler to accept float quantities

Frontend:
- Add handleGroupAssign() for equal split among all assignees
- Group mode now calculates: share = item.quantity / assignees.length
- All assignees get updated when anyone joins/leaves

CSS:
- Compact 3-level layout (name+price on same row)
- Smaller avatars (36px), badges, and steppers
- Reduced padding throughout for mobile efficiency

+++ backend/collaborative_session.py
@@ -138,7 +138,7 @@ def update_assignment(
     session_id: str,
     participant_id: str,
     item_id: str,
-    quantity: int,
+    quantity: float,  # Float to support fractional assignments (e.g., 0.33 for 3-way split)
     is_assigned: bool,
     updated_by: str
 ) -> Dict[str, Any]:
